+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [4.17.8] - YYYY-MM-DD
 ### Changed
 * Disabled SimplePie's HTML sanitization.
+* Images without an extension can now be imported.
+* The image importing function now allows the image URL and local path to be changed via filters.
 
 ### Fixed
 * WooCommerce Product type dropdown and accompanying options disappear while WP RSS Aggregator is active.